scraper/requirements.txt
aiohttp>=3.9.0
asyncpg>=0.29.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
httpx>=0.27.0
python-dotenv>=1.0.0
rich>=13.0.0
click>=8.1.0
tqdm>=4.66.0
openai>=1.12.0
ddgs>=0.5.0
pydantic>=2.0.0
